refactor(PortalProvider): colocate `useUnique` hook
